Updated mux signatures

+++ tensorflow_quantum/core/qsim/mux.cc
@@ -30,7 +30,7 @@ limitations under the License.
 namespace tfq {
 namespace qsim {
 
-StateSpace* GetStateSpace(const int num_qubits, const int num_threads) {
+StateSpace* GetStateSpace(const uint64_t num_qubits, const uint64_t num_threads) {
   if (num_qubits <= 3) {
     return new StateSpaceSlow(num_qubits, num_threads);
   }
