+++ tensorflow_quantum/core/qsim/mux.h
@@ -23,7 +23,7 @@ limitations under the License.
 namespace tfq {
 namespace qsim {
 
-StateSpace* GetStateSpace(const int num_qubits, const int num_threads);
+StateSpace* GetStateSpace(const uint64_t num_qubits, const uint64_t num_threads);
 
 }  // namespace qsim
 }  // namespace tfq
